# Use normal arrays + rkyv for improved perf