--- a/parcel_sourcemap/Cargo.toml
+++ b/parcel_sourcemap/Cargo.toml
@@ -11,7 +11,6 @@ repository = "https://github.com/parcel-bundler/source-map"
 [dependencies]
 "vlq" = "0.5.1"
 "napi" = "1.6.1"
-"serde" = { version = "1.0.126", features = ["derive"] }
-"bincode" = "1.3.3"
+rkyv = "0.6.6"
 "js-sys" = "0.3"
 "wasm-bindgen" = "0.2"