This replaces bincode with [rkyv](https://github.com/djkoloski/rkyv), a very fast zero copy rust serialization/deserialization library. It's similar to flatbuffers but doesn't require a separate schema, and also faster.

This also removes the use of `BTreeMap` for storing line and column mappings. On profiles, inserting into the btree was taking a significant amount of time, and replacing with a flattened sorted array allowed for faster serialization/deserialization and didn't hurt offsetting too much. Also rkyv doesn't support serializing BTreeMap, so this was required in order to get the perf wins from that.

## Benchmarks

### BTreeMap

```
consume vlq mappings 130,782.8 opts/sec (mean: 0.008ms, stddev: 0.017ms, 5,000 samples)
consume 280,260.56 opts/sec (mean: 0.004ms, stddev: 0.011ms, 5,000 samples)
consume JS Mappings 122,756.29 opts/sec (mean: 0.008ms, stddev: 0.01ms, 5,000 samples)
addSourceMap 133,329.46 opts/sec (mean: 0.008ms, stddev: 0.014ms, 5,000 samples)
positive line offset 278,254.51 opts/sec (mean: 0.004ms, stddev: 0.012ms, 5,000 samples)
negative line offset 296,130.71 opts/sec (mean: 0.003ms, stddev: 0.007ms, 5,000 samples)
positive column offset 293,881.07 opts/sec (mean: 0.003ms, stddev: 0.007ms, 5,000 samples)
negative column offset 299,806.98 opts/sec (mean: 0.003ms, stddev: 0.008ms, 5,000 samples)
```

### Array

```
consume vlq mappings 156,300.09 opts/sec (mean: 0.006ms, stddev: 0.02ms, 5,000 samples)
consume 347,612.32 opts/sec (mean: 0.003ms, stddev: 0.009ms, 5,000 samples)
consume JS Mappings 151,189.25 opts/sec (mean: 0.007ms, stddev: 0.011ms, 5,000 samples)
addSourceMap 143,650.78 opts/sec (mean: 0.007ms, stddev: 0.011ms, 5,000 samples)
positive line offset 219,920.37 opts/sec (mean: 0.005ms, stddev: 0.009ms, 5,000 samples)
negative line offset 200,892.86 opts/sec (mean: 0.005ms, stddev: 0.011ms, 5,000 samples)
positive column offset 362,958.15 opts/sec (mean: 0.003ms, stddev: 0.004ms, 5,000 samples)
negative column offset 343,651.02 opts/sec (mean: 0.003ms, stddev: 0.009ms, 5,000 samples)
```

### Array + rkyv

```
consume vlq mappings 159,484.08 opts/sec (mean: 0.006ms, stddev: 0.02ms, 5,000 samples)
consume 743,793.16 opts/sec (mean: 0.001ms, stddev: 0.009ms, 5,000 samples)
consume JS Mappings 153,713.45 opts/sec (mean: 0.007ms, stddev: 0.01ms, 5,000 samples)
addSourceMap 226,616.87 opts/sec (mean: 0.004ms, stddev: 0.013ms, 5,000 samples)
positive line offset 297,294.78 opts/sec (mean: 0.003ms, stddev: 0.012ms, 5,000 samples)
negative line offset 324,412.19 opts/sec (mean: 0.003ms, stddev: 0.004ms, 5,000 samples)
positive column offset 695,686.24 opts/sec (mean: 0.001ms, stddev: 0.008ms, 5,000 samples)
negative column offset 706,491.6 opts/sec (mean: 0.001ms, stddev: 0.007ms, 5,000 samples)
```

### Overall

So ~2.5x faster for consuming buffer mappings, which was taking a majority of the time. On the esbuild benchmark this shaves off ~200ms from packaging, or ~20%.