@devongovett 2021-06-12
macOS crash appears to be a bug in rkyv when used with jemalloc - I didn't see it on Apple Silicon though. https://github.com/djkoloski/rkyv/pull/153