@devongovett 2021-06-12
Also, benchmarks are currently kinda broken because of https://github.com/napi-rs/napi-rs/issues/613