@DeMoorJasper 2021-06-13
Impressive results, not sure if we should use a vector though as the benchmark shows it's as slow as it was before the Rust rewrite due to using the exact same logic as the CPP version.

Maybe we can write something in between vector and BTreeMap ourselves that can be serialised that uses less memory and is faster to manipulate?