@devongovett 2021-06-13
I didn't update the benchmark due to the napi bug, but I think the modify ones and the append ones are kinda incorrect because they continue to modify the same source map over and over again, which means that each iteration is different rather than the same. I think we'd need to move the sourcemap object construction _inside_ the benchmark iteration rather than outside so that each iteration is the same. Unfortunately, this means the timing will be less accurate as it will be dominated by the construction rather than the actual modification. Maybe there's a way to do this setup before each iteration but only measure the time of the actual modification?