@devongovett 2021-06-27
@DeMoorJasper what do you want to do with this? I couldn't find a way to measure only the append/offset time without also measuring the source map construction time, so unfortunately the benchmark will be mostly measuring that if we update it to move the construction inside each iteration rather than once at the start. I do think the current results are not really correct though, since each iteration is doing different work as the source map gets bigger and bigger which is not very real world. In real world benchmarks within Parcel this does improve perf significantly.